use Coord implementation in robot new fucniotns
