+++ coord.py
@@ -1,0 +1,7 @@
+class Coord:
+    def __init__(self, x, y):
+        self.x = x
+        self.y = y
+
+    def __str__(self):
+        return "("+str(self.x)+","+str(self.y)+")"
